ft(grid): add interfaces file
 - add interfaces file

[Delivers #167453425]
