+++ package.json
@@ -21,7 +21,6 @@
   "homepage": "https://github.com/jaodur/puzzled#readme",
   "devDependencies": {
     "@types/graphql": "^14.2.1",
-    "@types/history": "^4.7.2",
     "@types/react": "^16.8.20",
     "@types/react-dom": "^16.8.4",
     "@types/react-router-dom": "^4.3.4",
@@ -50,7 +49,6 @@
   "dependencies": {
     "apollo-boost": "^0.4.3",
     "graphql": "^14.3.1",
-    "history": "^4.9.0",
     "react": "^16.8.6",
     "react-apollo": "^2.5.6",
     "react-dom": "^16.8.6",
