+++ puzzled_ui/components/commons/link.tsx
@@ -1,14 +1,8 @@
 import * as React from "react";
 import { Link } from 'react-router-dom';
+import { routeLinkInterface } from '../interfaces'
 
-interface  routeLink {
-    link: string,
-    component: JSX.Element,
-    styleClass: string
-}
-
-
-function RouterLink({ link, component, styleClass }: routeLink) {
+function RouterLink({ link, component, styleClass }: routeLinkInterface) {
     return (
         <Link to={link} className={styleClass}>{component}</Link>
     )
