--- puzzled_ui/components/interface.ts
@@ -1,15 +1,0 @@
-interface sudokuGridInterface {
-    num: number
-}
-
-interface eventInterface {
-    target: any
-}
-
-interface  routeLinkInterface {
-    link: string,
-    component: JSX.Element,
-    styleClass: string
-}
-
-export { sudokuGridInterface, eventInterface }
