+++ puzzled_ui/components/interfaces.ts
@@ -1,0 +1,34 @@
+interface gridInterface {
+    num: number
+}
+
+interface eventInterface {
+    target: any
+}
+
+interface  routeLinkInterface {
+    link: string,
+    component: JSX.Element,
+    styleClass: string
+}
+
+interface gridRowInterface {
+    numItems: number,
+    sudokuGridClass: string
+}
+interface keyInterface {
+    key: string
+}
+
+interface gameIntroInterface {
+    gameClass: string
+}
+
+interface footerInterface {
+    footerClass: string
+}
+
+export {
+    gridInterface, eventInterface, routeLinkInterface, gridRowInterface, keyInterface, gameIntroInterface ,
+    footerInterface
+}
